+++ codex-rs/core/config.schema.json
@@ -1242,8 +1242,8 @@
       "default": null,
       "description": "Preferred backend for storing CLI auth credentials. file (default): Use a file in the Codex home directory. keyring: Use an OS-specific keyring service. auto: Use the keyring if available, otherwise use a file."
     },
-    "command_attribution": {
-      "description": "Optional command attribution text for commit message co-author trailers.\n\nSet to an empty string to disable automatic command attribution.",
+    "commit_attribution": {
+      "description": "Optional commit attribution text for commit message co-author trailers.\n\nSet to an empty string to disable automatic commit attribution.",
       "type": "string"
     },
     "compact_prompt": {
