+++ codex-rs/core/src/codex.rs
@@ -16,7 +16,7 @@ use crate::agent::agent_status_from_event;
 use crate::analytics_client::AnalyticsEventsClient;
 use crate::analytics_client::build_track_events_context;
 use crate::apps::render_apps_section;
-use crate::command_attribution::commit_message_trailer_instruction;
+use crate::commit_attribution::commit_message_trailer_instruction;
 use crate::compact;
 use crate::compact::run_inline_auto_compact_task;
 use crate::compact::should_use_remote_compact_task;
@@ -2430,7 +2430,7 @@ impl Session {
         }
         if turn_context.features.enabled(Feature::CodexGitCommit)
             && let Some(commit_message_instruction) = commit_message_trailer_instruction(
-                turn_context.config.command_attribution.as_deref(),
+                turn_context.config.commit_attribution.as_deref(),
             )
         {
             items.push(DeveloperInstructions::new(commit_message_instruction).into());
