+++ codex-rs/core/src/lib.rs
@@ -20,8 +20,8 @@ pub use codex_thread::CodexThread;
 pub use codex_thread::ThreadConfigSnapshot;
 mod agent;
 mod codex_delegate;
-mod command_attribution;
 mod command_canonicalization;
+mod commit_attribution;
 pub mod config;
 pub mod config_loader;
 pub mod connectors;
